# openapi.yaml
openapi: "3.0.3"
info:
  title: swagger sample
  version: "1.0"
servers:
  - url: http://localhost:3000
paths:
  '/health':
    get:
     summary: health check
     responses:
       '200':
        description: healty
        content:
          text/plain:
            schema:
              type: string
              example: server is healthy

  '/todo':
    post:
      summary: post todo
      requestBody:
          description: todo params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POSTTODORequest"
                required:
                  - subject
      responses:
        '200':
          description: return a created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TODOResponse"
        '400':
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        '405':
          description: method not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"

  '/todos':
    get:
      summary: get all todos
      responses:
        '200':
          description: return all todos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TODOsResponse"
        '500':
          description: internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StandardError"

components:
  schemas:
    TODO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        subject:
          type: string
        subscription:
          type: string
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    POSTTODORequest:
      type: object
      properties:
        subject:
          type: string
        subscription:
          type: string
    TODOResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        subject:
          type: string
        subscription:
          type: string
        createdAt:
          type: string
          format: date
        updatedAt:
          type: string
          format: date
    TODOsResponse:
      type: array
      items:
        $ref: "#/components/schemas/TODOResponse"
    StandardError:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string